Fixes an issue causing duplicate parameters in a single query to produce an invalid query.
When ApplyParameter find a parameter to replace the loop breaks leaving any subsequent uses of the parameter in question as the parameter key rather than the value.
Remove the break to ensure that the loop iterates over the entire query, allowing the method to find and replace all of the instances of a parameter.